input: touchscreen: fts: added touch key functionality
The ST Microelectronics FTM3BD56 touch screen device has a key
functinality to simulate the "phone" and the "back" button from
Android phone devices.

This patch adds this functionality as boards like TM2E are
using it instead of having separate device for the touch key.

The patch adds in the device tree a boolean property
"touch-key-connected" that, if present, enables the touch
property in the device and in the driver.

The driver, itself, when "touch-key-connected" present, generates
an input event that reports actions on the "phone" and on the
"back" key:

when the phone button is pressed or released:

	type  EV_KEY
	code  KEY_PHONE (169)
	value 1/0 (pressed or released)

when the back button is pressed or released:

	type  EV_KEY
	code  KEY_BACK (158)
	value 1/0 (pressed or released)

When discovering the deice, look for event name "sec_touchkey".

Signed-off-by: Andi Shyti <andi.shyti@samsung.com>