drm/exynos: decon: use deferred Trigger mode change
In case of DECON Display, the controller should work with SW trigger mode
even default mode is HW trigger because trigger mode should be changed
on PSR mode of Panel device - there might be some Panel not supported
for the PSR mode.

So this patch makes SW trigger mode to be used until dpms going off
to on again.

Signed-off-by: Inki Dae <inki.dae@samsung.com>