gpu: Generic graphic memory compression core
This patch introduces a frontend of GMC (graphic memory compression)
subsystem, which includes:
	- storage & compression facilities (based on zpool and crypto comp API);
	- user interface for triggering compression and decompression,
	  reading statistics.

The backend should be implemented inside a GPU driver.

NOTE:
These changes are temporary, there is no need of such subsystem in the
mainline kernel. The only purpose of the changes is to provide a
frontend for convinient and common usage by Midgard (Utgard)
platform GMC backends.

Change-Id: Ic2a9eba18e7c931fba1cb8843bad3727a5b09415
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>
Signed-off-by: Alexander Yashchenko <a.yashchenko@samsung.com>