gpu/mali: gmc: Remove alloc during invalidation
When region (alloc) is removed, the gmc handles associated with
this allocation are deleted. The old logic was to allocate new
clean page and pass it back to mali code which will add this
page to kbase_mem_pool (kbase_mem_pool_free_pages).
There are two problems with this logic:
- alloc_page overhead
- alloc page may fail and we need to consider the fault case

Instead of this we can just put zero values to phy_pages, which were
compressed at the moment of invalidation. These zeroed phy_pages
would be skipped in kbase_mem_pool_free_pages:
...
         for (i = 0; i < nr_pages; i++) {
                 if (unlikely(!pages[i]))
                         continue;
...
As a consequence, when compressed alloc is removed the pages wouldn't
be added to kbase_mem_pool. But
- If the compression was initiated, very likely, that reclaimable
mem pool is already dropped, so there is no need in low memory
conditions to add this pages to mali page cache (kbase_mem_pool)
- When we will have enough memory, the mali page cache will be
automatically filled up by not compressed allocations.

In other words there is no need in alloc_page on invalidation.

Change-Id: I898c7b8ab1f78914a5d73cc68da5c30c08c4d0cc
Signed-off-by: Alexander Yashchenko <a.yashchenko@samsung.com>