gpu/mali: gmc: guarantee mm_struct lifetime
This commit introduces code that manipulates a reference counter of mm_struct
(mm_count). The reference counter should be got to guarantee that the structure
won't be destroyed during the reclaim process.

Like in KSM, futexes and other kernel components operating with this reference
counter it is incremented under the task_lock critical section. The counter is
decremented with a call to mmdrop() which does atomic decrement and fetch and in
0-case destroys the mm_struct.

The counter mm_users cannot be used in spite of its name. Its nature is to count
a number of processes this mm_struct belongs to.

Change-Id: I0624f5c0f0acea2b03f5be4e6361f5c3585fa042
[ initial implementation ]
Signed-off-by: Alexander Yashchenko <a.yashchenko@samsung.com>
[ fix incorrect reference drop, new commit message ]
Signed-off-by: Sergei Rogachev <s.rogachev@samsung.com>